refactored api service classes
