+++ src/main/java/api_connection/FXRateService.java
@@ -1,10 +1,10 @@
-package api_connection.daily_service;
+package api_connection;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.Query;
 
-public interface DailyFXService {
+public interface FXRateService {
      String api_key = "556BNVOJG7TA17H2";
 
     /* gets 100 days worth of data
@@ -12,5 +12,14 @@ public interface DailyFXService {
      */
     @GET("query?function=FX_DAILY&from_symbol=USD&apikey=" + api_key)
     Call<TimeSeries> getDailyFXData(@Query("to_symbol") String to_symbol);
+
+    /**
+     *
+     *
+     * @param to_currency
+     * @return Realtime FX rate
+     */
+    @GET("query?function=CURRENCY_EXCHANGE_RATE&from_currency=USD&apikey=" + api_key)
+    Call<RealTimeFXRate> getRealTimeFXRate(@Query("to_currency") String to_currency);
 
 }
