+++ src/main/java/api_connection/RealTimeFXRate.java
@@ -1,4 +1,4 @@
-package api_connection.exchange_rates_service;
+package api_connection;
 
 import com.google.gson.annotations.SerializedName;
 import com.google.gson.internal.LinkedTreeMap;
